All unittests passes

+++ day11/adventofcode2019-day11/Day11Part1/AmplifierController.cs
@@ -5,31 +5,31 @@ namespace adventofcode2019_day11.Day11Part1
 {
     public class AmplifierController
     {
-        public int MaxThrusterSignal { get; private set; }
-        public List<int> Output { get; private set; }
+        public long MaxThrusterSignal { get; private set; }
+        public List<long> Output { get; private set; }
 
-        public static AmplifierController RunSequenceOnAmplifiers(int[] input, int seq1, int seq2, int seq3, int seq4, int seq5)
+        public static AmplifierController RunSequenceOnAmplifiers(long[] input, long seq1, long seq2, long seq3, long seq4, long seq5)
         {
             var result = new AmplifierController();
             result.Run(input, seq1, seq2, seq3, seq4, seq5);
             return result;
         }
 
-        public List<int[]> AllCombination(params int[] seq)
+        public List<long[]> AllCombination(params long[] seq)
         {
             var test = new FormPermut();
             test.PrnPermut(seq, 0, seq.Length - 1);
             return test.Permutations;
         }
 
-        private void Run(int[] code, int seq1, int seq2, int seq3, int seq4, int seq5)
+        private void Run(long[] code, long seq1, long seq2, long seq3, long seq4, long seq5)
         {
-            int? maxThrusterSignal = null;
+            long? maxThrusterSignal = null;
             var seqCombinations = new AmplifierController().AllCombination(seq1, seq2, seq3, seq4, seq5);
 
             foreach (var seq in seqCombinations)
             {
-                var resultAmp1 = IntCodeComputer.ProcessWithUserInput(Clone(code), new int[] { seq[0], 0 });
+                var resultAmp1 = IntCodeComputer.ProcessWithUserInput(Clone(code), new long[] { seq[0], 0 });
                 var resultAmp2 = IntCodeComputer.ProcessWithUserInput(Clone(code), SetInput(seq[1], resultAmp1.Output));
                 var resultAmp3 = IntCodeComputer.ProcessWithUserInput(Clone(code), SetInput(seq[2], resultAmp2.Output));
                 var resultAmp4 = IntCodeComputer.ProcessWithUserInput(Clone(code), SetInput(seq[3], resultAmp3.Output));
@@ -55,39 +55,39 @@ namespace adventofcode2019_day11.Day11Part1
         }
 
 
-        private int[] SetInput(int seq, List<int> output)
+        private long[] SetInput(long seq, List<long> output)
         {
-            var input = new List<int> { seq };
+            var input = new List<long> { seq };
             if (output != null)
                 input.AddRange(output);
 
             return input.ToArray();
         }
 
-        private int[] Clone(int[] input)
+        private long[] Clone(long[] input)
         {
-            return (int[])input.Clone();
+            return (long[])input.Clone();
         }
     }
     public class FormPermut
     {
-        public List<int[]> Permutations { get; private set; }
+        public List<long[]> Permutations { get; private set; }
         public FormPermut()
         {
-            Permutations = new List<int[]>();
+            Permutations = new List<long[]>();
         }
-        public void SwapTwoNumber(ref int a, ref int b)
+        public void SwapTwoNumber(ref long a, ref long b)
         {
-            int temp = a;
+            long temp = a;
             a = b;
             b = temp;
         }
-        public void PrnPermut(int[] list, int k, int m)
+        public void PrnPermut(long[] list, long k, long m)
         {
-            int i;
+            long i;
             if (k == m)
             {
-                var aPerm = new int[list.Length];
+                var aPerm = new long[list.Length];
                 for (i = 0; i <= m; i++)
                 {
                     aPerm[i] = list[i];
